+++ core/src/main/java/com/google/cloud/sql/core/CloudSqlInstance.java
@@ -25,7 +25,6 @@ import com.google.auth.oauth2.OAuth2Credentials;
 import com.google.cloud.sql.AuthType;
 import com.google.cloud.sql.CredentialFactory;
 import com.google.common.base.Throwables;
-import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningScheduledExecutorService;
@@ -42,7 +41,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.net.ssl.SSLSocket;
 
@@ -55,21 +53,19 @@ class CloudSqlInstance {
 
   private static final Logger logger = Logger.getLogger(CloudSqlInstance.class.getName());
   private static final String SQL_LOGIN_SCOPE = "https://www.googleapis.com/auth/sqlservice.login";
-  // defaultRefreshBuffer is the minimum amount of time for which a
-  // certificate must be valid to ensure the next refresh attempt has adequate
-  // time to complete.
-  private static final Duration DEFAULT_REFRESH_BUFFER = Duration.ofMinutes(4);
 
   private final ListeningScheduledExecutorService executor;
   private final SqlAdminApiFetcher apiFetcher;
   private final AuthType authType;
-  private final Optional<OAuth2Credentials> credentials;
+  private final Optional<OAuth2Credentials> iamAuthnCredentials;
   private final CloudSqlInstanceName instanceName;
   private final ListenableFuture<KeyPair> keyPair;
   private final Object instanceDataGuard = new Object();
   // Limit forced refreshes to 1 every minute.
   private final RateLimiter<Object> forcedRenewRateLimiter =
       RateLimiter.burstyBuilder(2, Duration.ofSeconds(30)).build();
+
+  private final RefreshCalculator refreshCalculator = new RefreshCalculator();
 
   @GuardedBy("instanceDataGuard")
   private ListenableFuture<InstanceData> currentInstanceData;
@@ -91,8 +87,7 @@ class CloudSqlInstance {
       AuthType authType,
       CredentialFactory tokenSourceFactory,
       ListeningScheduledExecutorService executor,
-      ListenableFuture<KeyPair> keyPair)
-      throws IOException, InterruptedException {
+      ListenableFuture<KeyPair> keyPair) {
     this.instanceName = new CloudSqlInstanceName(connectionName);
     this.apiFetcher = apiFetcher;
     this.authType = authType;
@@ -101,32 +96,15 @@ class CloudSqlInstance {
 
     if (authType == AuthType.IAM) {
       HttpRequestInitializer source = tokenSourceFactory.create();
-      this.credentials = Optional.of(parseCredentials(source));
-      this.credentials.get().refresh();
+      this.iamAuthnCredentials = Optional.of(parseCredentials(source));
     } else {
-      this.credentials = Optional.empty();
+      this.iamAuthnCredentials = Optional.empty();
     }
 
     synchronized (instanceDataGuard) {
       this.currentInstanceData = executor.submit(this::performRefresh);
       this.nextInstanceData = currentInstanceData;
     }
-  }
-
-  static long secondsUntilRefresh(Date expiration) {
-    Duration timeUntilExp = Duration.between(Instant.now(), expiration.toInstant());
-
-    if (timeUntilExp.compareTo(Duration.ofHours(1)) < 0) {
-      if (timeUntilExp.compareTo(DEFAULT_REFRESH_BUFFER) < 0) {
-        // If the time until the certificate expires is less the refresh buffer, schedule the
-        // refresh immediately
-        return 0;
-      }
-      // Otherwise schedule a refresh in (timeUntilExp - buffer) seconds
-      return timeUntilExp.minus(DEFAULT_REFRESH_BUFFER).getSeconds();
-    }
-    // If the time until the certificate expires is longer than an hour, return timeUntilExp//2
-    return timeUntilExp.dividedBy(2).getSeconds();
   }
 
   static GoogleCredentials getDownscopedCredentials(OAuth2Credentials credentials) {
@@ -162,8 +140,11 @@ class CloudSqlInstance {
         }
       };
     }
-
-    throw new RuntimeException("Not supporting credentials of type " + source.getClass().getName());
+    throw new RuntimeException(
+        String.format(
+            "[%s] Unable to connect via automatic IAM authentication: "
+                + "Unsupported credentials of type %s",
+            instanceName.getConnectionName(), source.getClass().getName()));
   }
 
   /**
@@ -244,66 +225,38 @@ class CloudSqlInstance {
     // To avoid unreasonable SQL Admin API usage, use a rate limit to throttle our usage.
     forcedRenewRateLimiter.acquirePermit();
 
-    ListenableFuture<InstanceData> refreshFuture;
-    if (authType == AuthType.IAM) {
-      if (credentials.isPresent()) {
-        GoogleCredentials downscopedCredentials = getDownscopedCredentials(credentials.get());
-        refreshFuture =
-            apiFetcher.getInstanceData(
-                instanceName, downscopedCredentials, AuthType.IAM, executor, keyPair);
-      } else {
-        throw new RuntimeException(
-            String.format(
-                "[%s] Unable to connect via automatic IAM authentication: Missing credentials.",
-                instanceName.getConnectionName()));
+    final GoogleCredentials credentials;
+    if (iamAuthnCredentials.isPresent()) {
+      credentials = getDownscopedCredentials(iamAuthnCredentials.get());
+    } else {
+      credentials = null;
+    }
+
+    try {
+      InstanceData data =
+          apiFetcher.getInstanceData(
+              this.instanceName, credentials, this.authType, executor, keyPair);
+
+      synchronized (instanceDataGuard) {
+        // update currentInstanceData with the most recent results
+        currentInstanceData = Futures.immediateFuture(data);
+
+        // schedule a replacement before the SSLContext expires;
+        nextInstanceData =
+            executor.schedule(
+                this::performRefresh,
+                refreshCalculator.calculateSecondsUntilNextRefresh(
+                    Instant.now(), data.getExpiration().toInstant()),
+                TimeUnit.SECONDS);
       }
 
-    } else {
-      refreshFuture =
-          apiFetcher.getInstanceData(instanceName, null, AuthType.PASSWORD, executor, keyPair);
+      return data;
+    } catch (ExecutionException | InterruptedException e) {
+      synchronized (instanceDataGuard) {
+        nextInstanceData = executor.submit(this::performRefresh);
+      }
+      throw e;
     }
-    Futures.addCallback(
-        refreshFuture,
-        new FutureCallback<InstanceData>() {
-          @Override
-          public void onSuccess(InstanceData instanceData) {
-            synchronized (instanceDataGuard) {
-              // update currentInstanceData with the most recent results
-              currentInstanceData = refreshFuture;
-              // schedule a replacement before the SSLContext expires;
-              nextInstanceData =
-                  executor.schedule(
-                      () -> performRefresh(),
-                      secondsUntilRefresh(getInstanceData().getExpiration()),
-                      TimeUnit.SECONDS);
-            }
-          }
-
-          @Override
-          public void onFailure(Throwable t) {
-            logger.log(
-                Level.WARNING,
-                "An error occurred while performing refresh. Retrying immediately.",
-                t);
-            synchronized (instanceDataGuard) {
-              InstanceData instanceData = null;
-              try {
-                instanceData = getInstanceData();
-              } catch (Exception e) {
-                // this means the result was invalid
-              }
-              if (instanceData == null
-                  || instanceData.getExpiration().toInstant().isBefore(Instant.now())) {
-                // replace current if it is expired or invalid
-                currentInstanceData = refreshFuture;
-              }
-              nextInstanceData = executor.submit(() -> performRefresh());
-            }
-          }
-        },
-        executor);
-
-    return refreshFuture.get();
   }
 
   private Optional<Date> getTokenExpirationTime(Credential credentials) {